requirements.txt
Flask==2.0.3
Flask-SocketIO==5.1.1
PyQt5==5.15.6
requests==2.26.0
python-socketio==5.4.1
bcrypt==3.2.0
pycryptodome==3.14.1